% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climacell_dict.R
\name{dict_moonphase}
\alias{dict_moonphase}
\title{Moonphase Dictionary}
\usage{
dict_moonphase()
}
\value{
a tibble
}
\description{
this is a helper function that returns the moon phase tibble containing the moon phase codes (which are returned by Climacell API) and their appropriate description.
}
