% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climacell_dict.R
\name{dict_preciptype}
\alias{dict_preciptype}
\title{Precipitation Type Dictionary}
\usage{
dict_preciptype()
}
\value{
a tibble
}
\description{
this is a helper function that returns the precipitation type tibble containing the precipitation type codes (which are returned by Climacell API) and their appropriate description.
}
