% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climacell_dict.R
\name{dict_weathercode}
\alias{dict_weathercode}
\title{Weather Dictionary}
\usage{
dict_weathercode()
}
\value{
a tibble
}
\description{
this is a helper function that returns the weather code tibble containing the weather codes (which are returned by Climacell API) and their appropriate description.
}
