% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climacell_temperature.R
\name{climacell_temperature}
\alias{climacell_temperature}
\title{Get Temperature Readings from Climacell}
\usage{
climacell_temperature(
  api_key,
  lat,
  long,
  timestep,
  start_time = NULL,
  end_time = NULL
)
}
\arguments{
\item{api_key}{a 32 character string that represents the private API key for accessing the Climacell API}

\item{lat}{a numeric value representing the latitude of the location}

\item{long}{a numeric value representing the longitude of the location}

\item{timestep}{a 'step' value for the time. Valid values are (choose one): c('1d', '1h', '30m','15m','5m','1m','current')}

\item{start_time}{the start date and time of the query. Must include timezone. Valid syntax: YYYY-MM-DDTHH:MM:SS-HH:MM. This value cannot be more than 6 hours prior to the current local time. This argument is OPTIONAL if timestep is 'current'.}

\item{end_time}{the end date and time of the query. Must include timezone. Valid syntax: YYYY-MM-DDTHH:MM:SS-HH:MM. The maximum end date/time depends on the timestep value chosen. The Climacell API will only return the maximum allowable results if the end_time is beyond the API limits. This argument is OPTIONAL if you do not wish to 'narrow' your end results.}
}
\value{
a tibble with the variables temperature, wind chill (or real feel temperature), dewpoint, and humidity
}
\description{
\code{climacell_temperature} will call the Climacell API and return temperature related attributes such as temperature, temperatureApparent (i.e., wind chill or real feel temperature), dew point, and humidity. All values are in metric and the timestamps returned by the API are always in UTC. These are not converted to the local machine's timezone.
}
