% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climacell_wind.R
\name{climacell_wind}
\alias{climacell_wind}
\title{Temperature Readings from Climacell}
\usage{
climacell_wind(
  api_key,
  lat,
  long,
  timestep,
  start_time = NULL,
  end_time = NULL
)
}
\arguments{
\item{api_key}{a 32 character string that represents the private API key for accessing the Climacell API. If missing, the function attempts to call on the environment variable called "CLIMACELL_API".}

\item{lat}{a numeric value (or a string that can be coerced to numeric) representing the latitude of the location}

\item{long}{a numeric value (or a string that can be coerced to numeric) representing the longitude of the location}

\item{timestep}{a 'step' value for the time. Valid values are (choose one): c('1d', '1h', '30m','15m','5m','1m','current')}

\item{start_time}{the start date and time of the query in \href{https://en.wikipedia.org/wiki/ISO_8601}{ISO8601} format. Must include timezone. Valid syntax: YYYY-MM-DDTHH:MM:SS-HH:MM. This value cannot be more than 6 hours prior to the current local time. This argument is OPTIONAL if timestep is 'current'.}

\item{end_time}{the end date and time of the query in \href{https://en.wikipedia.org/wiki/ISO_8601}{ISO8601} format. Must include timezone. Valid syntax: YYYY-MM-DDTHH:MM:SS-HH:MM. The maximum end date/time depends on the timestep value chosen. The Climacell API will only return the maximum allowable results if the end_time is beyond the API limits. This argument is OPTIONAL if you do not wish to 'narrow' your end results.}
}
\value{
a tibble
}
\description{
Temperature Readings from Climacell
}
